LayoutTests/platform/mac/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-2-expected.txt

⨀

⨁

⨃

⨅

⨇

⨈

⨉

⨊

⫼

⫿

⨀

⨁

⨃

⨅

⨇

⨈

⨉

⨊

⫼

⫿


FAIL Test that non-stretchy largeops are bigger in display mode. assert_greater_than: width of largeop '⫼' expected a number greater than 13.28125 but got 13.265625
